# First round of perf improvements for tiktoken