… src/Microsoft.ML.Tokenizers/PreTokenizer/PreTokenizer.cs
… src/Microsoft.ML.Tokenizers/PreTokenizer/TikTokenPreTokenizer.cs
… src/Microsoft.ML.Tokenizers/Tokenizer.cs
… src/Microsoft.ML.Tokenizers/TokenizerResult.cs
… src/Microsoft.ML.Tokenizers/Utils/BytePairEncoder.cs
… src/Microsoft.ML.Tokenizers/Utils/IListExtensions.cs