Before:
| Method              | Mean    | Allocated |
|-------------------- |--------:|----------:|
| CountTokensCached   | 3.677 s |   4.82 GB |
| CountTokensUncached | 2.309 s |   3.03 GB |

After:
| Method              | Mean    | Allocated |
|-------------------- |--------:|----------:|
| CountTokensCached   | 2.545 s | 637.63 MB |
| CountTokensUncached | 1.627 s | 408.34 MB |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using Microsoft.ML.Tokenizers;

BenchmarkSwitcher.FromAssembly(typeof(Tests).Assembly).Run(args);

[MemoryDiagnoser(false)]
[HideColumns("Job", "Error", "StdDev", "Median", "RatioSD")]
public class Tests
{
    private Tokenizer _tokenizer;
    private string[] _tests;

    [GlobalSetup]
    public async Task Setup()
    {
        _tokenizer = await Tokenizer.CreateByModelNameAsync("gpt-3.5-turbo");
        using HttpClient client = new HttpClient();
        string text = string.Concat(Enumerable.Repeat(Poem, 8));
        _tests = new string[8192]; // LruCache size
        for (int i = 0; i < _tests.Length; i++)
        {
            _tests[i] = text.Substring(0, text.Length - i);
        }
    }

    [Benchmark]
    public int CountTokensCached()
    {
        int sum = 0;
        for (int i = 0; i < _tests.Length; i++)
        {
            sum += _tokenizer.CountTokens(_tests[0]); // reuse same input each time
        }
        return sum;
    }

    [Benchmark]
    public int CountTokensUncached()
    {
        int sum = 0;
        for (int i = 0; i < _tests.Length; i++)
        {
            sum += _tokenizer.CountTokens(_tests[i]); // change the input to defeat the cache
        }
        return sum;
    }

    private const string Poem = """
        **Paws of Joy**

        In the morning's tender light,
        When dew-kissed grass awaits the sun,
        There stirs a creature, full of might,
        A friend whose loyalty is never undone.

        **The Dog**, with eyes like galaxies,
        Wags its tail, a metronome of glee,
        Its heart a map of boundless territories,
        Guiding us through life's vast sea.

        **Furry sentinels**, guardians of our hearth,
        They chase their tails in playful mirth,
        Their barks a symphony of love and merriment,
        Echoing through the quiet moments we've spent.

        **Nose to ground**, they follow scents,
        Unraveling mysteries with fervent intent,
        From squirrel trails to forgotten dreams,
        They lead us to places we've never seen.

        **Golden retrievers** with hearts of gold,
        **Dachshunds** with determination untold,
        **Greyhounds** racing against the wind,
        Each breed a chapter in the story they've pinned.

        **Labradors** dive into lakes with glee,
        **Chihuahuas** strut like tiny royalty,
        **Huskies** howl at the moon's silver glow,
        And **puppies**, oh sweet puppies, steal the show.

        Their eyes speak of trust, unwavering and true,
        Their fur holds secrets whispered by the dew,
        In their presence, worries seem to fade,
        As they teach us the art of living unafraid.

        So here's to the dogs, our steadfast friends,
        Who mend our hearts and heal life's bends,
        May their tails forever wag, their noses explore,
        For in their love, we find solace evermore.
        """;
}
```