@tarekgh 2024-02-18
> 

Would this better handled in the csproj? include this file only when we target net8.0.

---
Refers to: src/Microsoft.ML.Tokenizers/AssemblyInfo.cs:8 in ed88215. [](commit_id = ed88215736a510fca410327e968233f2daa9d009, deletion_comment = False)